fix_sqlalchemy_relationships.py
#!/usr/bin/env python3
"""
Fix SQLAlchemy Relationship Issues
"""
import os

def check_user_model_relationships():
    """Check relationships in User model"""
    print("🔍 Checking User Model Relationships")
    print("=" * 34)
    
    user_file = 'app/models/user.py'
    
    try:
        with open(user_file, 'r', encoding='utf-8') as f:
            content = f.read()
        
        print("✅ Found user model")
        
        # Look for relationship definitions
        lines = content.split('\n')
        relationships = []
        
        for i, line in enumerate(lines):
            if 'db.relationship' in line or 'relationship(' in line:
                relationships.append(f"Line {i+1}: {line.strip()}")
        
        print("Relationships found:")
        for rel in relationships:
            print(f"  {rel}")
        
        # Check for Email references
        if "'Email'" in content or '"Email"' in content:
            print("❌ Found string references to 'Email' - these cause issues")
        else:
            print("✅ No problematic Email references found")
        
        return content
        
    except Exception as e:
        print(f"❌ Error checking user model: {e}")
        return ""

def fix_user_model():
    """Fix the User model to use proper string references"""
    print("\n🔧 Fixing User Model")
    print("=" * 18)
    
    user_file = 'app/models/user.py'
    backup_file = 'app/models/user_backup.py'
    
    try:
        with open(user_file, 'r', encoding='utf-8') as f:
            content = f.read()
        
        # Backup current file
        with open(backup_file, 'w', encoding='utf-8') as f:
            f.write(content)
        
        print("✅ Backed up user model")
        
        # Create fixed user model
        fixed_user_model = '''"""
User model for AI Email Assistant
"""
from . import db
from datetime import datetime
from werkzeug.security import generate_password_hash, check_password_hash

class User(db.Model):
    """User model for authentication and profile data"""
    __tablename__ = 'users'
    
    id = db.Column(db.Integer, primary_key=True)
    
    # Basic user info
    email = db.Column(db.String(120), unique=True, nullable=False, index=True)
    display_name = db.Column(db.String(100), nullable=True)
    
    # Microsoft Azure AD info
    azure_id = db.Column(db.String(100), unique=True, nullable=True, index=True)
    azure_tenant_id = db.Column(db.String(100), nullable=True)
    
    # Authentication tokens (encrypted)
    access_token_hash = db.Column(db.Text, nullable=True)
    refresh_token_hash = db.Column(db.Text, nullable=True)
    token_expires_at = db.Column(db.DateTime, nullable=True)
    
    # User preferences
    timezone = db.Column(db.String(50), nullable=True, default='UTC')
    language = db.Column(db.String(10), nullable=True, default='en')
    
    # Account status
    is_active = db.Column(db.Boolean, nullable=False, default=True)
    is_admin = db.Column(db.Boolean, nullable=False, default=False)
    
    # Timestamps
    created_at = db.Column(db.DateTime, nullable=False, default=datetime.utcnow)
    updated_at = db.Column(db.DateTime, nullable=False, default=datetime.utcnow, onupdate=datetime.utcnow)
    last_login = db.Column(db.DateTime, nullable=True)
    last_sync = db.Column(db.DateTime, nullable=True)
    
    def __init__(self, **kwargs):
        """Initialize user with default values"""
        super(User, self).__init__(**kwargs)
    
    def __repr__(self):
        return f'<User {self.email}>'
    
    def to_dict(self):
        """Convert user to dictionary"""
        return {
            'id': self.id,
            'email': self.email,
            'display_name': self.display_name,
            'azure_id': self.azure_id,
            'timezone': self.timezone,
            'language': self.language,
            'is_active': self.is_active,
            'created_at': self.created_at.isoformat() if self.created_at else None,
            'last_login': self.last_login.isoformat() if self.last_login else None,
            'last_sync': self.last_sync.isoformat() if self.last_sync else None
        }
    
    @property
    def is_authenticated(self):
        """Check if user is authenticated"""
        return self.access_token_hash is not None
    
    @property 
    def has_valid_token(self):
        """Check if user has valid access token"""
        if not self.access_token_hash:
            return False
        if not self.token_expires_at:
            return True  # Assume valid if no expiry set
        return self.token_expires_at > datetime.utcnow()
'''
        
        with open(user_file, 'w', encoding='utf-8') as f:
            f.write(fixed_user_model)
        
        print("✅ Fixed user model (removed problematic relationships)")
        return True
        
    except Exception as e:
        print(f"❌ Error fixing user model: {e}")
        return False

def fix_email_model_relationships():
    """Fix Email model relationships"""
    print("\n🔧 Fixing Email Model Relationships")
    print("=" * 33)
    
    email_file = 'app/models/email.py'
    
    try:
        with open(email_file, 'r', encoding='utf-8') as f:
            content = f.read()
        
        # Check if it has relationship with User
        if 'db.relationship' in content and 'User' in content:
            print("❌ Found relationship that might cause issues")
            
            # Remove the relationship line for now
            lines = content.split('\n')
            fixed_lines = []
            
            for line in lines:
                if 'db.relationship' in line and ('User' in line or 'user' in line):
                    print(f"Removing: {line.strip()}")
                    continue
                else:
                    fixed_lines.append(line)
            
            fixed_content = '\n'.join(fixed_lines)
            
            with open(email_file, 'w', encoding='utf-8') as f:
                f.write(fixed_content)
            
            print("✅ Removed problematic relationships from email model")
        else:
            print("✅ Email model relationships look fine")
        
        return True
        
    except Exception as e:
        print(f"❌ Error fixing email model: {e}")
        return False

def create_clean_models_init():
    """Create a clean models/__init__.py without relationships"""
    print("\n🔧 Creating Clean Models Init")
    print("=" * 28)
    
    init_file = 'app/models/__init__.py'
    
    clean_init = '''"""
Models package for AI Email Assistant
"""
from flask_sqlalchemy import SQLAlchemy

# Initialize SQLAlchemy instance
db = SQLAlchemy()

# Import models without relationships to avoid circular references
from .user import User
from .email import Email
from .chat import ChatMessage

__all__ = ['db', 'User', 'Email', 'ChatMessage']
'''
    
    try:
        with open(init_file, 'w', encoding='utf-8') as f:
            f.write(clean_init)
        
        print("✅ Created clean models init")
        return True
        
    except Exception as e:
        print(f"❌ Error creating clean init: {e}")
        return False

def create_relationship_test():
    """Create a test to verify relationships work"""
    print("\n🧪 Creating Relationship Test")
    print("=" * 26)
    
    test_script = '''#!/usr/bin/env python3
"""
Test SQLAlchemy Relationships
"""
import os
import sys

def test_relationships():
    """Test if models can be imported without relationship errors"""
    print("🧪 Testing SQLAlchemy Relationships")
    print("=" * 33)
    
    os.environ['USE_DOCKER_CONFIG'] = 'true'
    sys.path.append('.')
    
    try:
        print("1. Importing models...")
        from app.models import db, User, Email, ChatMessage
        print("✅ All models imported successfully")
        
        print("2. Creating app...")
        from app import create_app
        app = create_app()
        print("✅ App created")
        
        print("3. Testing app context...")
        with app.app_context():
            print("✅ App context active")
            
            print("4. Creating tables...")
            db.create_all()
            print("✅ Tables created successfully")
            
            print("5. Testing model creation...")
            
            # Test user creation
            test_user = User(
                email="test@example.com",
                display_name="Test User",
                azure_id="test-123"
            )
            print("✅ User object created")
            
            # Test email creation
            test_email = Email(
                user_id=1,  # Will be set properly when user is saved
                message_id="test-message-123",
                subject="Test Email",
                sender_email="sender@example.com",
                body_text="Test email content"
            )
            print("✅ Email object created")
            
            print("\\n🎉 SUCCESS: All models work without relationship errors!")
            print("\\n🚀 You can now start the app: python docker_run.py")
            
        return True
        
    except Exception as e:
        print(f"\\n❌ FAILED: {e}")
        import traceback
        traceback.print_exc()
        return False

if __name__ == "__main__":
    test_relationships()
'''
    
    with open('test_relationships.py', 'w', encoding='utf-8') as f:
        f.write(test_script)
    
    print("✅ Created test_relationships.py")

def main():
    """Main function"""
    print("🔧 Fix SQLAlchemy Relationship Issues")
    print("=" * 36)
    print("🐛 Error: expression 'Email' failed to locate a name")
    print("This is a relationship definition issue between User and Email models")
    print()
    
    # Check current relationships
    check_user_model_relationships()
    
    # Fix user model
    user_fixed = fix_user_model()
    
    # Fix email model relationships
    email_fixed = fix_email_model_relationships()
    
    # Create clean models init
    init_fixed = create_clean_models_init()
    
    # Create test
    create_relationship_test()
    
    if user_fixed and email_fixed and init_fixed:
        print(f"\n🎉 RELATIONSHIP ISSUES FIXED!")
        print(f"=" * 28)
        
        print(f"\n🧪 Test the fix:")
        print(f"   python test_relationships.py")
        
        print(f"\n🚀 If test passes:")
        print(f"   python docker_run.py")
        
        print(f"\n✅ What was fixed:")
        print(f"   - Removed problematic relationships from User model")
        print(f"   - Cleaned up Email model relationships")
        print(f"   - Simple models without circular references")
        print(f"   - Clean imports that SQLAlchemy can handle")
        
        print(f"\n🎯 Models now have:")
        print(f"   - User model with all user fields")
        print(f"   - Email model with content fields (body_text, body_html, body_preview)")
        print(f"   - Chat model for AI conversations")
        print(f"   - No circular relationship dependencies")
        
        print(f"\n📧 After app starts:")
        print(f"   1. Fresh database with proper schema")
        print(f"   2. Open enhanced_sync_test.html")
        print(f"   3. Run enhanced sync")
        print(f"   4. Visit http://localhost:5000/emails/267")
        print(f"   5. Should show full email content!")
        
    else:
        print(f"\n❌ Some fixes failed")

if __name__ == "__main__":
    main()