requirements.txt
# Core Flask dependencies
Flask==3.0.0
Flask-SocketIO==5.3.6
Flask-CORS==4.0.0
Flask-Session==0.5.0

# Microsoft Graph API
msgraph-core
msgraph-sdk
azure-identity
msal

# Ollama integration
ollama==0.1.7
requests==2.31.0

# Vector database and embeddings
chromadb==0.4.18
sentence-transformers==2.2.2
numpy==1.24.3
faiss-cpu==1.7.4

# Database
SQLAlchemy==2.0.23
Flask-SQLAlchemy==3.1.1
alembic==1.12.1

# Authentication and security
PyJWT==2.8.0
cryptography
keyring==24.3.0

# Email processing
email-validator==2.1.0
python-dateutil==2.8.2
beautifulsoup4==4.12.2
markdownify==0.11.6

# Utilities
python-dotenv==1.0.0
schedule==1.2.0
apscheduler==3.10.4
click==8.1.7

# Development and testing
pytest==7.4.3
pytest-flask==1.3.0
black==23.11.0
flake8==6.1.0

# Windows-specific
pywin32==306
#winsound

# Optional: Better logging
structlog==23.2.0